+++ src/Products/PloneMeeting/browser/views.py
@@ -48,6 +48,7 @@ from Products.PloneMeeting.utils import _base_extra_expr_ctx
 from Products.PloneMeeting.utils import _itemNumber_to_storedItemNumber
 from Products.PloneMeeting.utils import _storedItemNumber_to_itemNumber
 from Products.PloneMeeting.utils import get_annexes
+from Products.PloneMeeting.utils import get_dx_field
 from Products.PloneMeeting.utils import get_dx_widget
 from Products.PloneMeeting.utils import get_person_from_userid
 from Products.PloneMeeting.utils import signatureNotAlone
@@ -455,6 +456,39 @@ class BaseDGHV(object):
 
     def __init__(self, context, request):
         self.printed_scan_id_barcode = []
+
+    def _print_special_value(self, field_name, empty_marker=u'', **kwargs):
+        """Overridable method to manage some specific usecases for self.print_value."""
+        return None
+
+    def print_value(self, field_name, empty_marker=u'', **kwargs):
+        """Convenient method that print more or less everything."""
+        # special handling for some values
+        value = self._print_special_value(field_name, empty_marker, **kwargs)
+        if value is None:
+            # get attribute in schema
+            value = getattr(self.real_context, field_name)
+            field = get_dx_field(self.real_context, field_name)
+            class_name = field.__class__.__name__
+            # Datetime
+            if class_name == 'Datetime':
+                if 'custom_format' not in kwargs:
+                    kwargs['custom_format'] = '%-d %B %Y'
+                value = self.display_date(date=getattr(self.real_context, field_name), **kwargs)
+            # RichText
+            elif class_name == 'RichText':
+                value = value and value.output
+                if value:
+                    value = self.printXhtml(self.context, xhtmlContents=value, **kwargs)
+            # List/Choice
+            elif getattr(field, 'vocabulary', None):
+                value = self.display_voc(field_name, **kwargs)
+
+        # if a p_empty_marker is given and no value, use it
+        # it may be "???" or "xxx" for example
+        if not value:
+            value = empty_marker
+        return value
 
     def image_orientation(self, image):
         """Compute image orientation, if orientation is landscape, we rotate
@@ -1717,6 +1751,12 @@ class FolderDocumentGenerationHelperView(ATDocumentGenerationHelperView, BaseDGH
 class MeetingDocumentGenerationHelperView(DXDocumentGenerationHelperView, FolderDocumentGenerationHelperView):
     """ """
 
+    def _print_special_value(self, field_name, empty_marker='', **kwargs):
+        """Manage 'place' manually."""
+        # 'place' may be stored in 'place' or 'place_other'
+        if field_name == 'place':
+            return self.real_context.get_place()
+
 
 class ItemDocumentGenerationHelperView(ATDocumentGenerationHelperView, BaseDGHV):
     """ """
@@ -1732,7 +1772,7 @@ class ItemDocumentGenerationHelperView(ATDocumentGenerationHelperView, BaseDGHV)
             result['public_deliberation_decided'] = self.print_public_deliberation_decided()
         return result
 
-    def print_meeting_date(self, returnDateTime=False, noMeetingMarker='-', unrestricted=True):
+    def print_meeting_date(self, returnDateTime=False, noMeetingMarker='-', unrestricted=True, **kwargs):
         """Print meeting date, manage fact that item is not linked to a meeting,
            in this case p_noMeetingMarker is returned.
            If p_returnDateTime is True, it returns the meeting date DateTime,
@@ -1746,9 +1786,11 @@ class ItemDocumentGenerationHelperView(ATDocumentGenerationHelperView, BaseDGHV)
 
         if returnDateTime:
             return meeting.date
-        return meeting.Title()
+        else:
+            dghv = self.getDGHV(meeting)
+            return dghv.print_value("date", **kwargs)
 
-    def print_preferred_meeting_date(self, returnDateTime=False, noMeetingMarker='-', unrestricted=True):
+    def print_preferred_meeting_date(self, returnDateTime=False, noMeetingMarker='-', unrestricted=True, **kwargs):
         """
         Print preferred meeting date, manage fact that item has no preferred meeting date
         :param returnDateTime if True, returns the preferred meeting date DateTime, otherwise it returns the
@@ -1764,7 +1806,9 @@ class ItemDocumentGenerationHelperView(ATDocumentGenerationHelperView, BaseDGHV)
 
         if returnDateTime:
             return preferred_meeting.date
-        return preferred_meeting.Title()
+        else:
+            dghv = self.getDGHV(preferred_meeting)
+            return dghv.print_value("date", **kwargs)
 
     def print_in_and_out_attendees(
             self,